Add explicit Safe Haskell annotations
Add explicit Safe/Trustworthy annotations to every module.

Move the writeByteString function from Blaze.ByteString.Builder.ByteString
to Blaze.ByteString.Builder.Internal.Write because it uses rather unsafe
functions compared to the rest of the module and fits better in the
unsafe context of the latter module. This makes it possible to mark the former
module Safe. There is no impact on the public API (if you consider the Internal
modules private).